# swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing fresh products and orders",
        "title": "Meli Fresh Products API",
        "contact": {
            "name": "Bootcampers GO - W5"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sections": {
            "get": {
                "description": "Fetches all sections available in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Retrieve all sections",
                "responses": {
                    "200": {
                        "description": "List of sections",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.Section"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new section with the provided details on the request body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Create a new section",
                "parameters": [
                    {
                        "description": "Section Create Request",
                        "name": "section",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RequestSectionJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Section",
                        "schema": {
                            "$ref": "#/definitions/internal.Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "409": {
                        "description": "Section with given section number already registered\" or \"Warehouse not found\" or \"Product-type not found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "422": {
                        "description": "Couldn't parse section",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/api/v1/sections/report-products": {
            "get": {
                "description": "Fetches a report of products available in a section or across all sections",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Retrieve a report of products in a section",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Section ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report of products in sections",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ResponseReportProd"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Section not found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/api/v1/sections/{id}": {
            "get": {
                "description": "Fetches the section based on the provided section ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Retrieve a section by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section data",
                        "schema": {
                            "$ref": "#/definitions/internal.Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Section not found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a section identified by its Id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Delete a section",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Section not found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a section with the provided Id and data on the request body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Section"
                ],
                "summary": "Update an existing section",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated section data",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Section",
                        "schema": {
                            "$ref": "#/definitions/internal.Section"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Section not found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "409": {
                        "description": "Section with given section number already registered",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.RequestSectionJSON": {
            "type": "object",
            "properties": {
                "current_capacity": {
                    "type": "integer"
                },
                "current_temperature": {
                    "type": "number"
                },
                "maximum_capacity": {
                    "type": "integer"
                },
                "minimum_capacity": {
                    "type": "integer"
                },
                "minimum_temperature": {
                    "type": "number"
                },
                "product_type_id": {
                    "type": "integer"
                },
                "section_number": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
        "handler.ResponseReportProd": {
            "type": "object",
            "properties": {
                "products_count": {
                    "type": "integer"
                },
                "section_id": {
                    "type": "integer"
                },
                "section_number": {
                    "type": "integer"
                }
            }
        },
...
        "internal.Section": {
            "type": "object",
            "properties": {
                "current_capacity": {
                    "type": "integer"
                },
                "current_temperature": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "maximum_capacity": {
                    "type": "integer"
                },
                "minimum_capacity": {
                    "type": "integer"
                },
                "minimum_temperature": {
                    "type": "number"
                },
                "product_type_id": {
                    "type": "integer"
                },
                "section_number": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "rest_err.Causes": {
            "description": "Structure representing the causes of an error.",
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field associated with the error cause.\n@json\n@jsonTag field",
                    "type": "string",
                    "example": "name"
                },
                "message": {
                    "description": "Error message describing the cause.\n@json\n@jsonTag message",
                    "type": "string",
                    "example": "name is required"
                }
            }
        },
        "rest_err.RestErr": {
            "description": "Structure for describing why the error occurred",
            "type": "object",
            "properties": {
                "causes": {
                    "description": "Error causes.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_err.Causes"
                    }
                },
                "code": {
                    "description": "Error code.",
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "description": "Error description.",
                    "type": "string",
                    "example": "internal_server_error"
                },
                "message": {
                    "description": "Error message.",
                    "type": "string",
                    "example": "error trying to process request"
                }
            }
        }
    }
}